update instructions in the full-stack tutorial to include HttpClientManager.kt and move the file to the network folder
