+++ codeSnippets/snippets/tutorial-full-stack-task-manager/README.md
@@ -9,7 +9,7 @@ by following the steps in the [Full-stack development with Kotlin Multiplatform
 Because it's not possible to make calls to `0.0.0.0` or `localhost` from code running on an Android Virtual Device or the
 iPhone simulator, you need to specify a host address for the HTTP client.
 
-1. Navigate to the `TaskApi.kt` file in `composeApp/src/commonMain/kotlin/org/example/ktor/network`.
+1. Navigate to the `HttpClientManager.kt` file in `composeApp/src/commonMain/kotlin/org/example/ktor/network`.
 2. In the `defaultRequest` function replace the value of the `host` property with the IP address of your machine.
 
 ## Run the server
@@ -20,7 +20,7 @@ Use the following command to start the server:
 ./gradlew :server:run
 ```
 
-The server will then run at [http://0.0.0.0:8080/](). To test the endpoinds, navigate to the following URLs:
+The server will then run at [http://0.0.0.0:8080/](). To test the endpoints, navigate to the following URLs:
 
 - [http://0.0.0.0:8080/tasks](http://0.0.0.0:8080/tasks) to see a complete list of tasks in JSON format.
 - [http://0.0.0.0:8080/tasks/byPriority/Medium](http://0.0.0.0:8080/tasks/byPriority/Medium) to see tasks filtered
