+++ codeSnippets/snippets/tutorial-full-stack-task-manager/composeApp/src/commonMain/kotlin/org/example/ktor/network/HttpClientManager.kt
@@ -1,4 +1,4 @@
-package org.example.ktor.core.network
+package org.example.ktor.network
 
 import io.ktor.client.HttpClient
 import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
